requirements/base.txt
asgiref==3.8.1
async-timeout==4.0.3
cfgv==3.4.0
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-modeltranslation==0.18.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.3
gunicorn==21.2.0
identify==2.5.35
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
virtualenv==20.25.1
